Moved unit test inside sudoku folder
