+++ Tests/Sudoku/GameStateMemoryTests.cs
@@ -2,7 +2,7 @@
 using XenobiaSoft.Sudoku;
 using XenobiaSoft.Sudoku.GameState;
 
-namespace UnitTests;
+namespace UnitTests.Sudoku;
 
 public class GameStateMemoryTests : BaseTestByAbstraction<GameStateMemory, IGameStateMemory>
 {
